python/requirements.txt
beautifulsoup4==4.7.1
Cython==0.29.24
language-tool-python==2.6.1
nltk==3.6.2
numpy==1.19.5
progressbar2==4.0.0